+++ kpi-tracker/resources/schema.sql
@@ -35,8 +35,13 @@ CREATE INDEX IF NOT EXISTS blocks_timestamp ON blocks (timestamp);
 CREATE TABLE IF NOT EXISTS paydays (
   block INT8 PRIMARY KEY REFERENCES blocks(id) ON DELETE RESTRICT ON UPDATE RESTRICT,
   total_equity_capital INT8 NOT NULL,
-  total_delegated_stake INT8 NOT NULL,
+  total_passively_delegated INT8 NOT NULL,
+  total_actively_delegated INT8 NOT NULL,
   num_bakers INT8 NOT NULL,
+  num_open_bakers INT8 NOT NULL,
+  num_closed_bakers INT8 NOT NULL,
+  num_open_delegation_recipients INT8 NOT NULL,
+  num_closed_for_new_delegation_recipients INT8 NOT NULL,
   num_finalizers INT8 NOT NULL,
   num_delegators INT8 NOT NULL
 );